requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==3.7.1
async-timeout==5.0.1
attrs==25.3.0
azure-core==1.33.0
azure-cosmos==4.7.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
gradio==3.28.1
gradio_client==1.8.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.1
idna==3.10
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
linkify-it-py==2.0.3
markdown-it-py==2.2.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
multidict==6.3.2
narwhals==1.33.0
numpy==2.2.4
openai==0.27.8
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pdfminer.six==20221105
pdfplumber==0.9.0
pillow==11.1.0
propcache==0.3.1
protobuf==5.29.4
pyarrow==19.0.1
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
pydeck==0.9.1
pydub==0.25.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
semantic-version==2.10.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.1
streamlit==1.44.1
tabulate==0.9.0
tenacity==8.2.2
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
uc-micro-py==1.0.3
urllib3==2.3.0
uvicorn==0.34.0
Wand==0.6.13
watchdog==6.0.0
websockets==15.0.1
yarl==1.19.0
rank_bm25 
nltk==3.8.1
tiktoken
azure-ai-textanalytics
